+++ jdaviz/configs/cubeviz/plugins/moment_maps/moment_maps.vue
@@ -1,3 +1,5 @@
+<script setup>
+</script>
 <template>
   <j-tray-plugin
     :config="config"
@@ -56,12 +58,19 @@
       :items="continuum_subset_items"
       v-model:selected="continuum_subset_selected"
       :show_if_single_entry="true"
-      :rules="[() => continuum_subset_selected!==spectral_subset_selected || 'Must not match line selection.']"
       label="Continuum"
-      api_hint="plg.continuum_subset ="
+      api_hint="plg.continuum ="
       :api_hints_enabled="api_hints_enabled"
       hint="Select spectral region that defines the continuum."
     />
+
+    <j-flex-row v-if="continuum_subset_selected===spectral_subset_selected">
+      <v-alert type="error" style="width: 100%; margin-left: 12px; margin-right: 12px;">
+        <span class="v-messages v-messages__message text--secondary">
+            Cannot use the same region to define both the spectral region and the continuum.
+        </span>
+      </v-alert>
+    </j-flex-row>
 
     <plugin-dataset-select
       v-if="continuum_subset_selected !== 'None'"
@@ -162,7 +171,7 @@
       action_label="Calculate"
       action_tooltip="Calculate moment map"
       :action_spinner="spinner"
-      :action_disabled="n_moment > 0 && output_unit_selected !== 'Spectral Unit' && reference_wavelength === 0"
+      :action_disabled="continuum_subset_selected===spectral_subset_selected || (n_moment > 0 && output_unit_selected !== 'Spectral Unit' && reference_wavelength === 0)"
       add_results_api_hint = 'plg.add_results'
       action_api_hint='plg.calculate_moment(add_data=True)'
       :api_hints_enabled="api_hints_enabled"