+++ CHANGES.rst
@@ -144,6 +144,8 @@ Bug Fixes
 
 - Fix error when typing while renaming data or subset. [#4359]
 
+
+
 Mosviz
 ^^^^^^
 
@@ -172,6 +174,8 @@ Other Changes and Additions
 
 Bug Fixes
 ---------
+
+- Fix a bug in Line Analysis that looks for a (possibly) non-existent mask when getting the continuum [#4406]
 
 Mosviz
 ^^^^^^